% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_anomaly_score.R
\name{compute_anomaly_score}
\alias{compute_anomaly_score}
\title{Compute anomaly scores}
\usage{
compute_anomaly_score(obj, x, window_size = 1L, ...)
}
\arguments{
\item{obj}{object}

\item{x}{data.frame to compute anomaly scores}

\item{window_size}{integer.}

\item{...}{additional arguments}
}
\description{
Compute anomaly scores
}
