% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMRFmix.R
\name{GMRFmix}
\alias{GMRFmix}
\title{Gaussian Markov random field mixtures}
\usage{
GMRFmix(x, pi, m, A, alpha = rep(1, K), max_iter = 400, tol = 0.01,
  verbose = TRUE)
}
\description{
Gaussian Markov random field mixtures
}
