% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_multivariate_data}
\alias{plot_multivariate_data}
\title{Plot multivariate data}
\usage{
plot_multivariate_data(df, label = NULL, order_by = index(df),
  guide_title = NULL, fix_scale = FALSE, point_size = 1L)
}
\arguments{
\item{df}{data.frame of multivariate data}

\item{label}{data.frame of label for each variables. Or vector of label for each observation.}

\item{order_by}{vector. An x-axis of plots.}

\item{guide_title}{character.}

\item{fix_scale}{logical.}

\item{point_size}{integer. Point size.}
}
\value{
ggplot2 object
}
\description{
Plot multivariate data
}
\examples{
library(sGMRFmix)

test_data <- generate_test_data()
test_label <- generate_test_labels()

plot_multivariate_data(test_data)
plot_multivariate_data(test_data, test_label)

}
