% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sGMRFmix.R
\name{sGMRFmix}
\alias{sGMRFmix}
\title{Sparse Gaussian Markov Random Field Mixtures}
\usage{
sGMRFmix(x, K, rho, m0 = rep(0, M), lambda0 = 1, pi_threshold = 1/K/100,
  max_iter = 400, tol = 0.01, verbose = TRUE)
}
\description{
Sparse Gaussian Markov Random Field Mixtures
}
