% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseGaussMix.R
\name{sparseGaussMix}
\alias{sparseGaussMix}
\title{Sparse mixture of sparse graphical models}
\usage{
sparseGaussMix(x, K, rho, m0 = rep(0, M), lambda0 = 1, max_iter = 400,
  tol = 0.01, verbose = TRUE)
}
\description{
Sparse mixture of sparse graphical models
}
